requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.26.4
matplotlib>=3.5.0
opencv-python>=4.5.5
scikit-learn>=1.0.2
pillow>=9.0.0
tqdm>=4.64.0
albumentations>=1.3.0
